% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/masonmd.R
\name{classify.nmd}
\alias{classify.nmd}
\title{Classify NMD-elicit mutations}
\usage{
classify.nmd(gene_id, ref = 37, mut_start, mut_end, ref_nt = "-",
  mut_nt = "-")
}
\arguments{
\item{gene_id}{numeric; gene Entrez ID}

\item{ref}{numeric; reference genome/NCBI build; \code{ref = 37} if the mutation is called by
build37/hg19, or \code{ref = 36} if the mutation is called by build36/hg18}

\item{mut_start}{numeric; the absolution start position of the mutations on chromosome on + strand}

\item{mut_end}{numeric; the absolution end position of the mutations on chromosome on + strand}

\item{ref_nt}{charactor; the reference nucleotides of the mutations on + strand;
\code{ref_nt = "-"} for insertion}

\item{mut_nt}{charactor; the mutation nucleotides on + strand; \code{mut_nt = "-"} for deletion}
}
\value{
A list containing eight entries:
\itemize{
\item{\code{mut_nmd}} {whether the mutation is NMD-elicit.}
\item{\code{note}} {extra explanation of the classification results.}
\item{\code{wt_nmd}} {whether the wildtype CDS harbours PTC that is NMD-elicit.}
\item{\code{PTC.Stop}} {the position of PTC.}
\item{\code{have.ptc}} {if the mutated sequence harbours a PTC.}
\item{\code{mutseq_length}} {length of the mutated coding sequences}
\item{\code{last_exon_exon_junction}} {relative position of last exon-exon junction}
\item{\code{n.exon}} {number of exons in the gene}
}
}
\description{
\code{classify.nmd} can detect the premature stop codon (PTC) and predict NMD-elicit mutations
based on three rules.
}
\details{
classify the mutations into NMD or non-NMD, and return the position of premature stop codon in CDS
}
\examples{
library(masonmd)
# an example of NMD-escape mutation from TCGA
classify.nmd(gene_id = 13, ref = 37, mut_start = 151545640, mut_end = 151545640,
ref_nt = "G",mut_nt = "T")
# an example of NMD-elicit mutation from TCGA
classify.nmd(gene_id = 2, ref = 37, mut_start = 9221429, mut_end = 9221429,
ref_nt = "G", mut_nt = "A")
}

