% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/masonmd.R
\name{findORF}
\alias{findORF}
\title{Find ORF}
\usage{
findORF(start_pos, k, stop_pos)
}
\arguments{
\item{start_pos}{all possible start positions}

\item{k}{minimal length for ORF}

\item{stop_pos}{all possible stop positions}
}
\value{
the start positions, stop positions and lengths of all posible ORFs
}
\description{
Find open reading frame (ORF)
}
\details{
Find the ORF in '+' strand
}

