% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/masonmd.R
\name{findStartPos}
\alias{findStartPos}
\title{Find start position}
\usage{
findStartPos(start_codons = c("ATA", "ATG"), seq)
}
\arguments{
\item{start_codons}{a character vector, the stop codons used; default is TGA, TAA and TAG.}

\item{seq}{a character vector, sequence.}
}
\value{
potential start codons
}
\description{
Find all start codons in the sequence
}
\details{
Find start codon in the sequence
}

