% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/masonmd.R
\name{findStopPos}
\alias{findStopPos}
\title{Find stop position}
\usage{
findStopPos(stop_codons = c("TGA", "TAA", "TAG"), seq)
}
\arguments{
\item{stop_codons}{a character vector, the stop codons used; default is TGA, TAA and TAG.}

\item{seq}{a character vector, sequence.}
}
\value{
{stop_pos} {found all stop positions regardless of the frames}
}
\description{
Find all stop codons in the sequence
}
\details{
Find all stop codons in the sequence
}

