% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/masonmd.R
\name{get.longest.cds}
\alias{get.longest.cds}
\title{Get information of longest isoform for given gene}
\usage{
get.longest.cds(gene_id, ref)
}
\arguments{
\item{gene_id}{numeric; Entrez gene ID}

\item{numeric;}{reference genome/NCBI build; \code{ref = 37} if the mutation is called by
build37/hg19, or \code{ref = 36} if the mutation is called by build36/hg18}
}
\value{
A list containing four entries:
\itemize{
\item{\code{gene}} {gene Entrez ID; same with the input}
\item{\code{txid}} {transcript ID of the longest isoform}
\item{\code{seq}} {charactor; coding sequence of the transcript}
\item{\code{tx}} {infomation of the transcript in form of Txdb}
}
}
\description{
\code{sum} returns the sum of all the values present in its arguments.
}
\details{
Get the longest cds if the gene ID is given
}
\examples{
library(masonmd)
get.longest.cds(gene_id = 13, ref = 37)
}

