% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/masonmd.R
\name{get_mut_exjun}
\alias{get_mut_exjun}
\title{Get relative positions of exon-exon junctions in mutated gene}
\usage{
get_mut_exjun(wt_exjun, mut_start, mut_end, mut_nt)
}
\arguments{
\item{wt_exjun}{a string that consists of 0 or 1 to indicates the positions of exon-exon junctions in
the coding sequence;}

\item{mut_start}{position of mutation start in chromosome}

\item{mut_end}{position of mutation end in chromosome}

\item{mut_nt}{nucleotides after alteration}
}
\value{
A list containing two entries:
\itemize{
\item{\code{mut_exjun}} {charactor that consists of 0 and 1; 1 represents the position of junctions}
\item{\code{note}} {extra note}
}
}
\description{
\code{get_mut_exjun} refers relative positions of exon-exon junctions in mutated gene.
}
\details{
Indicate the positions of exon-exon junctions in the mutated sequence based on the mutation
information and the positions of exon-exon junctions in wildtype sequence.
}

