% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/masonmd.R
\name{get_mut_seq_exjun}
\alias{get_mut_seq_exjun}
\title{Get mutated coding sequence and junctions}
\usage{
get_mut_seq_exjun(wt_exjun, wt_seq, mut_start, mut_end, mut_nt, ref_nt, note,
  txid)
}
\arguments{
\item{wt_exjun}{a string that consists of 0 or 1 to indicates the positions of exon-exon junctions in
the coding sequence;}

\item{wt_seq}{wildtype coding sequence}

\item{mut_start}{position of mutation start in chromosome}

\item{mut_end}{position of mutation end in chromosome}

\item{mut_nt}{nucleotides after alteration}

\item{ref_nt}{the nucleotides on the reference genome at the mutated position; the nucleotides before
mutation}

\item{note}{extra notes}

\item{txid}{transcript ID}
}
\value{
A list containing three entries:
\itemize{
\item{\code{mut_exjun}} {numeric string with 0 and 1; 1 represents the position of junctions}
\item{\code{note}} {extra note}
}
}
\description{
Get mutated coding sequence and junctions
}
\details{
Get the mutated sequence and junctions
}

