% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/masonmd.R
\name{get_tx_seq}
\alias{get_tx_seq}
\title{Get coding sequence of a transcript}
\usage{
get_tx_seq(tx_pos, ncbi_build)
}
\arguments{
\item{tx_pos}{The positional information of transcipt from TxDB}

\item{ncbi_build}{NCBI_build}
}
\value{
A list containing two entries:
\itemize{
\item{\code{wt_pos}} {numeric string; the absolute positions on chromosomes for each nucleotide on the coding sequence}
\item{\code{exonJunction}} {numeric string with 0 and 1; 1 represents the position of junctions}
}
}
\description{
Get positions and junction for the coding sequence of a transcript
}
\details{
Get the absolute positions on chromosomes for each nucleotide on the coding sequence for
a given transcript; and indicate the exon-exon junctions in the coding sequence
}

