% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/masonmd.R
\name{last.exexjun.pos}
\alias{last.exexjun.pos}
\title{Get position of last exon-exon junction}
\usage{
last.exexjun.pos(mut_exjun)
}
\arguments{
\item{mut_exjun}{a numeric vector that identifies the positions of exon-exon junctions
in the coding sequence. It is the output of the function \code{get_mut_seq_exjun} in masonmd package.}
}
\value{
A list containing two entries:
\itemize{
\item{\code{last_exon_exon_junction}} {numeric; position of last exon-exon junction.}
\item{\code{note}} {note on weather the mutation affects splice sites or the gene has only one exon.}
}
}
\description{
\code{last.exexjun.pos} gets position of last exon-exon junction.
}
\details{
Get position of last exon-exon junction.
}

