% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/masonmd.R
\name{mut.info}
\alias{mut.info}
\title{Get the information of mutated sequence}
\usage{
mut.info(gene_id, ref, mut_start, mut_end, ref_nt, mut_nt)
}
\arguments{
\item{gene_id}{numeric; gene Entrez ID}

\item{ref}{numeric; reference genome/NCBI build; \code{ref = 37} if the mutation is called by
build37/hg19, or \code{ref = 36} if the mutation is called by build36/hg18}

\item{mut_start}{numeric; the absolution start position of the mutations on chromosome on + strand}

\item{mut_end}{numeric; the absolution end position of the mutations on chromosome on + strand}

\item{ref_nt}{charactor; the reference nucleotides of the mutations on + strand;
\code{ref_nt = "-"} for insertion}

\item{mut_nt}{charactor; the mutation nucleotides on + strand; \code{mut_nt = "-"} for deletion}
}
\value{
charactor vector containing five entries:
\itemize{
\item{\code{mut_seq}} {charactor; coding sequence of the mutated gene.}
\item{\code{last.exexjun}} {numeric; position of last exon-exon junction.}
\item{\code{note}} {extra explanation.}
\item{\code{wt_seq}} {charactor; coding sequence of the wildtype gene.}
\item{\code{n.exon}} {numeric; number of exons in the gene.}
}
}
\description{
\code{mut.info} returns the sequences and positional infomation of a mutated gene.
}
\details{
Returns the sequences and positional infomation of a mutated gene.
}
\examples{
library(masonmd)
# an example of NMD-escape mutation from TCGA
mut.info(gene_id = 13, ref = 37, mut_start = 151545640, mut_end = 151545640,
ref_nt = "G",mut_nt = "T")
}

